fix(timestamp): Used `new Date().getTime()` instead of `performance.now()`
Changed `waitUntil` function to be compatible with NodeJS without the need for a polyfill.
Now it uses `new Date().getTime()` instead of `performance.now()`. This change was necessary to fix the `ReferenceError: performance is not defined` error when trying to run `waitUntil` function inside of a node script. This is because `performance` is only defined inside the browsers (which this module was originally created for).